port: wire ResourceRegistry/BiomeRegistry to JSON data, add textured tile/sprite rendering and TextRenderer
- BiomeDef/DataRecord/ResourceDef: mutable setters for System.Text.Json binding;
  ResourceDef field names now match resources.json envelope keys
- Bootstrap: build BiomeRegistry directly from deserialized BiomeDef list
  (replaces reflection-based dictionary copy) and construct ResourceRegistry
- WorldGen: sigmoid elevation curve so lowlands cluster and peaks stay rare;
  biome thresholds rescaled to 0..31 elevation; ResourcePlacer receives the
  ResourceRegistry and honors per-resource density and sprite key
- Render: PrimitiveBatch gains textured quads (sprite + terrain overlay) with
  color/textured mode flushing; TileRenderer projects all four tile corners
  with per-corner elevation and overlays biome terrain textures; SpriteRenderer
  loads/caches sprite textures from Assets/sprites
- TextRenderer + bundled Liberation Mono fonts; title screen uses it
- Constants: DomainWarpAmplitude 12 -> 4 to reduce biome smearing

+++ src/DontStarveRuneScape/World/Tile.cs
@@ -19,7 +19,7 @@ public sealed class ResourceNode
     public bool IsDepleted => Density <= 0f;
 
     // Additional properties for action system compatibility
-    public int XpReward => ResourceDef?.Xp ?? 0;
+    public int XpReward => (int)(ResourceDef?.Xp ?? 0);
     public string YieldItem => ResourceDef?.Id ?? string.Empty;
     public int YieldQuantity => ResourceDef?.Yield ?? 1;
     public bool RequiresTool => ResourceDef?.RequiresTool ?? false;
@@ -79,7 +79,7 @@ public sealed class ResourceNode
         LastHarvestTime = 0; // Will be set by world time
         TotalHarvests++;
 
-        return (def.Id, quantity, def.Xp);
+        return (def.Id, quantity, (int)def.Xp);
     }
 
     /// <summary>
